+++ Tests/SentryTests/SentryOptionsTest.m
@@ -205,6 +205,11 @@
     [self testBooleanField:@"enableLogs" defaultValue:NO];
 }
 
+- (void)testEnableMetrics
+{
+    [self testBooleanField:@"enableMetrics" defaultValue:NO];
+}
+
 - (void)testEnableAutoBreadcrumbTracking
 {
     [self testBooleanField:@"enableAutoBreadcrumbTracking"];